be: Rework job prepare/schedule
The pispbe_schedule_internal() function is meant to be called with the
hw_lock taken and realease it conditionally if it can schedule a job.

This 'called with lock, conditionally unlock' pattern causes the caller
to implement a cumbersome error handling and triggers a few
smatch/sparse warnings.

Split the job preparation and scheduling to two different parts, by
introducing a job_descriptor which is prepared by pispbe_prepare_job()
which needs to be called with the lock help, and scheduled by
pispbe_schedule() which is now in control of the lock/unlock sequences.

Signed-off-by: Jacopo Mondi <jacopo.mondi@ideasonboard.com>